requirements.txt
google-ads>=23.1.0
pandas>=2.1.0
pyyaml>=6.0
argparse
tqdm
